Create llm-python-file-embedding-by-line.py
The basic Python to read in a file line by line and generate the embeddings.  It outputs CSV format, 'filename,text,embedding'.  In this instance we'll manually edit the 'filename' to be any text we want returned when the embedding matches, in this use case it's a path on the server.